requirements.txt
langchain
langchain-community
langchain-openai
flask
flasgger
python-dotenv
aiohttp
playwright
browser-use
jira